+++ src/wallet/wallet.h
@@ -979,7 +979,7 @@ public:
         return m_pool_key_to_index;
     }
     /** Does the wallet have at least min_keys in the keypool? */
-    bool HasUnusedKeys(int min_keys) const;
+    bool HasUnusedKeys(size_t min_keys) const;
 
     std::set<std::set<CTxDestination>> GetAddressGroupings();
     std::map<CTxDestination, Amount> GetAddressBalances();
